moved all payload check functions to the payloadCheck library
