+++ src/lib/orionld/common/CMakeLists.txt
@@ -28,8 +28,6 @@ SET (SOURCES
     urnCheck.cpp
     linkCheck.cpp
     httpStatusCodeToOrionldErrorType.cpp
-    geoJsonTypeCheck.cpp
-    geoJsonCheck.cpp
     geoJsonCreate.cpp
     numberToDate.cpp
     orionldState.cpp
