+++ src/lib/orionld/serviceRoutines/orionldPatchRegistration.cpp
@@ -34,15 +34,15 @@ extern "C"
 #include "common/globals.h"                                     // parse8601Time
 #include "rest/ConnectionInfo.h"                                // ConnectionInfo
 
-#include "orionld/common/CHECK.h"                               // STRING_CHECK, ...
 #include "orionld/common/orionldState.h"                        // orionldState
 #include "orionld/common/orionldErrorResponse.h"                // orionldErrorResponseCreate
-#include "orionld/common/geoJsonCheck.h"                        // geoJsonCheck
+#include "orionld/common/urlCheck.h"                            // urlCheck
+#include "orionld/common/urnCheck.h"                            // urnCheck
 #include "orionld/context/orionldContextItemExpand.h"           // orionldContextItemExpand
 #include "orionld/context/orionldContextValueExpand.h"          // orionldContextValueExpand
 #include "orionld/context/orionldContextItemAlreadyExpanded.h"  // orionldContextItemAlreadyExpanded
 #include "orionld/db/dbConfiguration.h"                         // dbRegistrationGet, dbRegistrationReplace
-#include "orionld/payloadCheck/pcheckEntities.h"                // pcheckEntities
+#include "orionld/payloadCheck/pcheckRegistration.h"            // pcheckRegistration
 #include "orionld/serviceRoutines/orionldPatchRegistration.h"   // Own Interface
 
 
@@ -519,421 +519,6 @@ static bool ngsildRegistrationToAPIv1Datamodel(ConnectionInfo* ciP, KjNode* patc
 
 
 
-// -----------------------------------------------------------------------------
-//
-// orionldCheckGeoJsonGeometry -
-//
-static bool orionldCheckGeoJsonGeometry(ConnectionInfo* ciP, KjNode* geoAttributeP, const char* fieldName)
-{
-  return geoJsonCheck(ciP, geoAttributeP, NULL, NULL);
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// orionldCheckRegistrationInformation
-//
-// The "information" field can have only three members:
-//   * entities       (Mandatory JSON Array)
-//   * properties     (Optional JSON Array with strings)
-//   * relationships  (Optional JSON Array with strings)
-//
-// NOTE
-//   This function also expands ATTRIBUTE NAMES and ENTITY TYPES
-//
-static bool orionldCheckRegistrationInformation(ConnectionInfo* ciP, KjNode* informationP)
-{
-  KjNode* entitiesP = NULL;
-
-  for (KjNode* infoItemP = informationP->value.firstChildP; infoItemP != NULL; infoItemP = infoItemP->next)
-  {
-    if (strcmp(infoItemP->name, "entities") == 0)
-    {
-      DUPLICATE_CHECK(entitiesP, "entities", infoItemP);
-      if (pcheckEntities(ciP, entitiesP) == false)
-        return false;
-    }
-    else if (strcmp(infoItemP->name, "properties") == 0)
-    {
-      ARRAY_CHECK(infoItemP, "information[X]::properties");
-      EMPTY_ARRAY_CHECK(infoItemP, "information[X]::properties");
-      for (KjNode* propP = infoItemP->value.firstChildP; propP != NULL; propP = propP->next)
-      {
-        STRING_CHECK(propP, "information[X]::properties[X]");
-        EMPTY_STRING_CHECK(propP, "information[X]::properties[X]");
-        if (orionldContextItemAlreadyExpanded(propP->value.s) == false)
-          propP->value.s = orionldContextItemExpand(orionldState.contextP, propP->value.s, NULL, true, NULL);
-      }
-    }
-    else if (strcmp(infoItemP->name, "relationships") == 0)
-    {
-      ARRAY_CHECK(infoItemP, "information[X]::relationships");
-      EMPTY_ARRAY_CHECK(infoItemP, "information[X]::relationships");
-      for (KjNode* relP = infoItemP->value.firstChildP; relP != NULL; relP = relP->next)
-      {
-        STRING_CHECK(relP, "information[X]::relationships[X]");
-        EMPTY_STRING_CHECK(relP, "information[X]::relationships[X]");
-        if (orionldContextItemAlreadyExpanded(relP->value.s) == false)
-          relP->value.s = orionldContextItemExpand(orionldState.contextP, relP->value.s, NULL, true, NULL);
-      }
-    }
-    else
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for information[X]", infoItemP->name);
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// orionldCheckRegistrationInformationArray
-//
-static bool orionldCheckRegistrationInformationArray(ConnectionInfo* ciP, KjNode* informationArrayP, const char* fieldName)
-{
-  if (informationArrayP->value.firstChildP == NULL)  // Empty Array
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Empty Array", "information");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  for (KjNode* informationP = informationArrayP->value.firstChildP; informationArrayP != NULL; informationArrayP = informationArrayP->next)
-  {
-    OBJECT_CHECK(informationP, "information[X]");
-
-    if (orionldCheckRegistrationInformation(ciP, informationP) == false)
-      return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// orionldCheckObservationInterval -
-//
-static bool orionldCheckObservationInterval(ConnectionInfo* ciP, KjNode* timeIntervalNodeP, const char* fieldName)
-{
-  KjNode* startP = NULL;
-  KjNode* endP   = NULL;
-  int64_t start  = 0;
-  int64_t end    = 0;
-
-  for (KjNode* tiItemP = timeIntervalNodeP->value.firstChildP; tiItemP != NULL; tiItemP = tiItemP->next)
-  {
-    if (strcmp(tiItemP->name, "start") == 0)
-    {
-      DUPLICATE_CHECK(startP, "observationInterval::start", tiItemP);
-      STRING_CHECK(tiItemP, "observationInterval::start");
-      DATETIME_CHECK(tiItemP->value.s, start, "observationInterval::start");
-    }
-    else if (strcmp(tiItemP->name, "end") == 0)
-    {
-      DUPLICATE_CHECK(endP, "observationInterval::end", tiItemP);
-      STRING_CHECK(tiItemP, "observationInterval::end");
-      DATETIME_CHECK(tiItemP->value.s, end, "observationInterval::end");
-    }
-    else
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for observationInterval", tiItemP->name);
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-  }
-
-  if ((startP == NULL) && (endP == NULL))
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Empty Object", "observationInterval");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (startP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Missing mandatory field", "observationInterval::start");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (endP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Missing mandatory field", "observationInterval::end");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (start > end)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Inconsistent DateTimes", "observationInterval ends before it starts");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// orionldCheckManagementInterval -
-//
-static bool orionldCheckManagementInterval(ConnectionInfo* ciP, KjNode* timeIntervalNodeP, const char* fieldName)
-{
-  KjNode* startP = NULL;
-  KjNode* endP   = NULL;
-  int64_t start  = 0;
-  int64_t end    = 0;
-
-  for (KjNode* tiItemP = timeIntervalNodeP->value.firstChildP; tiItemP != NULL; tiItemP = tiItemP->next)
-  {
-    if (strcmp(tiItemP->name, "start") == 0)
-    {
-      DUPLICATE_CHECK(startP, "managementInterval::start", tiItemP);
-      STRING_CHECK(tiItemP, "managementInterval::start");
-      DATETIME_CHECK(tiItemP->value.s, start, "managementInterval::start");
-    }
-    else if (strcmp(tiItemP->name, "end") == 0)
-    {
-      DUPLICATE_CHECK(endP, "managementInterval::end", tiItemP);
-      STRING_CHECK(tiItemP, "managementInterval::end");
-      DATETIME_CHECK(tiItemP->value.s, end, "managementInterval::end");
-    }
-    else
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for managementInterval", tiItemP->name);
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-  }
-
-  if ((startP == NULL) && (endP == NULL))
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Empty Object", "managementInterval");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (startP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Missing mandatory field", "managementInterval::start");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (endP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Missing mandatory field", "managementInterval::end");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (start > end)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Inconsistent DateTimes", "managementInterval ends before it starts");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// registrationPayloadCheck -
-//
-// This function makes sure that the incoming payload is OK
-// - all types are as they should,
-// - the values with special needs are correct (if DataTikme for example)
-// - no field is duplicated
-// - etc
-//
-// However, there is ONE thing that cannot be checked here, as we still haven't queries the database
-// for the original registration - the one to be patched.
-// The problem is with registration properties.
-// A registration can have any property added to it, and with any name.
-// In this function we can check that the property isn't duplicated but, as this is a PATCH, the
-// property to be patched MUST ALREADY EXIST in the original registration, and this cannot be checked until
-// AFTER we extract the the original registration from the database.
-//
-// SO, what we do here is to separate the regiustration properties from the rest of the fields and keep thjat
-// as a KjNode tree, until we are able to perform this check.
-// The output parameter 'propertyTreeP' is used for this purpose.
-//
-static bool registrationPayloadCheck(ConnectionInfo* ciP, KjNode* registrationP, bool idCanBePresent, KjNode**  propertyTreeP)
-{
-  KjNode*  idP                   = NULL;
-  KjNode*  typeP                 = NULL;
-  KjNode*  nameP                 = NULL;
-  KjNode*  descriptionP          = NULL;
-  KjNode*  informationP          = NULL;
-  KjNode*  observationIntervalP  = NULL;
-  KjNode*  managementIntervalP   = NULL;
-  KjNode*  locationP             = NULL;
-  KjNode*  observationSpaceP     = NULL;
-  KjNode*  operationSpaceP       = NULL;
-  KjNode*  expiresP              = NULL;
-  KjNode*  endpointP             = NULL;
-  KjNode*  propertyTree          = kjObject(orionldState.kjsonP, NULL);  // Temp storage for properties
-  int64_t  dateTime;
-
-  if (registrationP->type != KjObject)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Registration", "The payload data for updating a registration must be a JSON Object");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-
-  //
-  // Loop over the entire registration (incoming payload data) and check all is OK
-  //
-  // To be able to easy and fast check that no registration properties are duplicated, the registration properties are
-  // removed from the tree and added to 'propertyTree'.
-  // Later, these registration properties must be put back!
-  //
-  // It's quicker to implement this like that, as there's no need to go over the entire path tree and also not necessary to look for >1 matches.
-  // When Property P1 is about to be added to 'propertyTree' for a second time, the duplication will be noticed and the error triggered.
-  //
-  KjNode* nodeP = registrationP->value.firstChildP;
-  KjNode* next;
-
-  while (nodeP != NULL)
-  {
-    next = nodeP->next;
-
-    if (strcmp(nodeP->name, "id") == 0)
-    {
-      if (idCanBePresent == false)
-      {
-        orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for Registration Update", "id");
-        ciP->httpStatusCode = SccBadRequest;
-        return false;
-      }
-
-      DUPLICATE_CHECK(idP, "id", nodeP);
-      STRING_CHECK(nodeP, "id");
-      URI_CHECK(nodeP, "id");
-    }
-    else if (strcmp(nodeP->name, "type") == 0)
-    {
-      DUPLICATE_CHECK(typeP, "type", nodeP);
-      STRING_CHECK(nodeP, "type");
-
-      if (strcmp(nodeP->value.s, "ContextSourceRegistration") != 0)
-      {
-        orionldErrorResponseCreate(OrionldBadRequestData, "Invalid value for type", nodeP->value.s);
-        ciP->httpStatusCode = SccBadRequest;
-        return false;
-      }
-    }
-    else if (strcmp(nodeP->name, "name") == 0)
-    {
-      DUPLICATE_CHECK(nameP, "name", nodeP);
-      STRING_CHECK(nodeP, "name");
-    }
-    else if (strcmp(nodeP->name, "description") == 0)
-    {
-      DUPLICATE_CHECK(descriptionP, "description", nodeP);
-      STRING_CHECK(nodeP, "description");
-    }
-    else if (strcmp(nodeP->name, "information") == 0)
-    {
-      DUPLICATE_CHECK(informationP, "information", nodeP);
-      ARRAY_CHECK(nodeP, "information");
-      if (orionldCheckRegistrationInformationArray(ciP, nodeP, "information") == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "observationInterval") == 0)
-    {
-      DUPLICATE_CHECK(observationIntervalP, "observationInterval", nodeP);
-      OBJECT_CHECK(nodeP, "observationInterval");
-      EMPTY_OBJECT_CHECK(nodeP, "observationInterval");
-      if (orionldCheckObservationInterval(ciP, nodeP, "observationInterval") == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "managementInterval") == 0)
-    {
-      DUPLICATE_CHECK(managementIntervalP, "managementInterval", nodeP);
-      OBJECT_CHECK(nodeP, "managementInterval");
-      if (orionldCheckManagementInterval(ciP, nodeP, "managementInterval") == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "location") == 0)
-    {
-      DUPLICATE_CHECK(locationP, "location", nodeP);
-      OBJECT_CHECK(nodeP, "location");
-
-      if (orionldCheckGeoJsonGeometry(ciP, locationP, "location") == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "observationSpace") == 0)
-    {
-      DUPLICATE_CHECK(observationSpaceP, "observationSpace", nodeP);
-      OBJECT_CHECK(nodeP, "observationSpace");
-      if (orionldCheckGeoJsonGeometry(ciP, nodeP, "observationSpace") == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "operationSpace") == 0)
-    {
-      DUPLICATE_CHECK(operationSpaceP, "operationSpace", nodeP);
-      OBJECT_CHECK(nodeP, "operationSpace");
-      if (orionldCheckGeoJsonGeometry(ciP, nodeP, "operationSpace") == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "expires") == 0)
-    {
-      DUPLICATE_CHECK(expiresP, "expires", nodeP);
-      STRING_CHECK(nodeP, "expires");
-      DATETIME_CHECK(expiresP->value.s, dateTime, "expires");
-    }
-    else if (strcmp(nodeP->name, "endpoint") == 0)
-    {
-      DUPLICATE_CHECK(endpointP, "endpoint", nodeP);
-      STRING_CHECK(nodeP, "endpoint");
-      URI_CHECK(nodeP, nodeP->name);
-    }
-    else
-    {
-      kjChildRemove(registrationP, nodeP);
-
-      //
-      // Add the Property to 'propertyTree'
-      //
-      // But, before adding, look it up.
-      // If already there, then we have a case of duplicate property and that is an error
-      //
-      if (kjLookup(propertyTree, nodeP->name) != NULL)
-      {
-        ciP->httpStatusCode = SccBadRequest;
-        orionldErrorResponseCreate(OrionldBadRequestData, "Duplicate Property in Registration Update", nodeP->name);
-        return false;
-      }
-
-      kjChildAdd(propertyTree, nodeP);
-    }
-
-    nodeP = next;
-  }
-
-  *propertyTreeP = propertyTree;  // These properties will be checked later, after getting the reg from DB
-
-  return true;
-}
-
-
-
 // ----------------------------------------------------------------------------
 //
 // ngsildRegistrationPatch -
@@ -986,9 +571,9 @@ bool orionldPatchRegistration(ConnectionInfo* ciP)
     return false;
   }
 
-  if (registrationPayloadCheck(ciP, orionldState.requestTree, false, &propertyTree) == false)
+  if (pcheckRegistration(ciP, orionldState.requestTree, false, &propertyTree) == false)
   {
-    LM_E(("registrationPayloadCheck FAILED"));
+    LM_E(("pcheckRegistration FAILED"));
     return false;
   }
