+++ src/lib/orionld/serviceRoutines/orionldPatchSubscription.cpp
@@ -1,6 +1,6 @@
 /*
 *
-* Copyright 2018 FIWARE Foundation e.V.
+* Copyright 2019 FIWARE Foundation e.V.
 *
 * This file is part of Orion-LD Context Broker.
 *
@@ -34,369 +34,13 @@ extern "C"
 
 #include "rest/ConnectionInfo.h"                                // ConnectionInfo
 
-#include "orionld/common/CHECK.h"                               // STRING_CHECK, ...
 #include "orionld/common/orionldState.h"                        // orionldState
 #include "orionld/common/orionldErrorResponse.h"                // orionldErrorResponseCreate
-#include "orionld/context/orionldContextItemExpand.h"           // orionldContextItemExpand
-#include "orionld/payloadCheck/pcheckEntities.h"                // pcheckEntities
+#include "orionld/common/urlCheck.h"                            // urlCheck
+#include "orionld/common/urnCheck.h"                            // urnCheck
+#include "orionld/payloadCheck/pcheckSubscription.h"            // pcheckSubscription
 #include "orionld/db/dbConfiguration.h"                         // dbSubscriptionGet
 #include "orionld/serviceRoutines/orionldPatchSubscription.h"   // Own Interface
-
-
-
-// -----------------------------------------------------------------------------
-//
-// ngsildCoordinatesToAPIv1Datamodel -
-//
-static bool ngsildCoordinatesToAPIv1Datamodel(ConnectionInfo* ciP, KjNode* coordinatesP, const char* fieldName, KjNode* geometryP)
-{
-  bool   isPoint = false;
-  char*  buf;
-
-  if (geometryP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Internal Error", "Unable to extract the geometry of a geoQ for coordinmate APIv1 fix");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  if (strcmp(geometryP->value.s, "Point") == 0)
-    isPoint = true;
-
-  if (isPoint)
-  {
-    // A point is an array ( [ 1, 2 ] ) in NGSI-LD, but in APIv1 database mode it is a string ( "1,2" )
-    int    coords = 0;
-    float  coordV[3];
-
-    for (KjNode* coordP = coordinatesP->value.firstChildP; coordP != NULL; coordP = coordP->next)
-    {
-      if (coordP->type == KjFloat)
-        coordV[coords] = coordP->value.f;
-      else
-        coordV[coords] = (float) coordP->value.i;
-
-      ++coords;
-    }
-
-    buf = kaAlloc(&orionldState.kalloc, 128);
-    if (coords == 2)
-      snprintf(buf, 128, "%f,%f", coordV[0], coordV[1]);
-    else
-      snprintf(buf, 128, "%f,%f,%f", coordV[0], coordV[1], coordV[2]);
-  }
-  else
-  {
-    buf = kaAlloc(&orionldState.kalloc, 1024);
-    kjRender(orionldState.kjsonP, coordinatesP, buf, 1024);
-  }
-
-  coordinatesP->name    = (char*) "coords";
-  coordinatesP->type    = KjString;
-  coordinatesP->value.s = buf;
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// pcheckGeoQ - FIXME: move to payloadCheck library
-//
-static bool pcheckGeoQ(ConnectionInfo* ciP, KjNode* geoqNodeP)
-{
-  //
-  // Render the coordinates and convert it into a string - for the NGSIv1 database model
-  //
-  KjNode* geometryP    = kjLookup(geoqNodeP, "geometry");
-  KjNode* coordinatesP = kjLookup(geoqNodeP, "coordinates");
-
-  if ((coordinatesP != NULL) && (coordinatesP->type == KjArray))
-  {
-    if (ngsildCoordinatesToAPIv1Datamodel(ciP, coordinatesP, "geoQ::coordinates", geometryP) == false)
-      return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// pcheckEndpoint - FIXME: move to payloadCheck library
-//
-bool pcheckEndpoint(ConnectionInfo* ciP, KjNode* endpointP)
-{
-  KjNode* uriP    = NULL;
-  KjNode* acceptP = NULL;
-
-  for (KjNode* epItemP = endpointP->value.firstChildP; epItemP != NULL; epItemP = epItemP->next)
-  {
-    if (strcmp(epItemP->name, "uri") == 0)
-    {
-      DUPLICATE_CHECK(uriP, "endpoint::uri", epItemP);
-      STRING_CHECK(uriP, "endpoint::uri");
-      URI_CHECK(uriP, "endpoint::uri");
-    }
-    else if (strcmp(epItemP->name, "accept") == 0)
-    {
-      DUPLICATE_CHECK(acceptP, "endpoint::accept", epItemP);
-      STRING_CHECK(acceptP, "endpoint::accept");
-      if ((strcmp(acceptP->value.s, "application/json") != 0) && (strcmp(acceptP->value.s, "application/ld+json") != 0))
-      {
-        orionldErrorResponseCreate(OrionldBadRequestData, "Unsupported Mime-type in 'accept'", epItemP->value.s);
-        ciP->httpStatusCode = SccBadRequest;
-        return false;
-      }
-    }
-    else
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for 'endpoint'", epItemP->name);
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-  }
-
-  if (uriP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Mandatory field missing in 'endpoint'", "uri");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// pcheckNotification - FIXME: move to payloadCheck library
-//
-bool pcheckNotification(ConnectionInfo* ciP, KjNode* notificationP)
-{
-  KjNode* attributesP = NULL;
-  KjNode* formatP     = NULL;
-  KjNode* endpointP   = NULL;
-
-  for (KjNode* nItemP = notificationP->value.firstChildP; nItemP != NULL; nItemP = nItemP->next)
-  {
-    if (strcmp(nItemP->name, "attributes") == 0)
-    {
-      DUPLICATE_CHECK(attributesP, "attributes", nItemP);
-      ARRAY_CHECK(nItemP, "attributes");
-
-      for (KjNode* attrP = nItemP->value.firstChildP; attrP != NULL; attrP = attrP->next)
-      {
-        STRING_CHECK(attrP, "attributes array item");
-        attrP->value.s = orionldContextItemExpand(orionldState.contextP, attrP->value.s, NULL, true, NULL);
-      }
-    }
-    else if (strcmp(nItemP->name, "format") == 0)
-    {
-      DUPLICATE_CHECK(formatP, "format", nItemP);
-      STRING_CHECK(formatP, "format");
-      if ((strcmp(formatP->value.s, "keyValues") != 0) && (strcmp(formatP->value.s, "normalized") != 0))
-      {
-        orionldErrorResponseCreate(OrionldBadRequestData, "Invalid value of 'format' (must be either 'keyValues' or 'normalized'", formatP->value.s);
-        ciP->httpStatusCode = SccBadRequest;
-        return false;
-      }
-    }
-    else if (strcmp(nItemP->name, "endpoint") == 0)
-    {
-      DUPLICATE_CHECK(endpointP, "endpoint", nItemP);
-      OBJECT_CHECK(endpointP, "endpoint");
-      if (pcheckEndpoint(ciP, endpointP) == false)
-        return false;
-    }
-    else if (strcmp(nItemP->name, "status") == 0)
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for notification", "'status' is read-only");
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-    else
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Invalid field for notification", nItemP->name);
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-  }
-
-  if (endpointP == NULL)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Mandatory field missing", "endpoint");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  return true;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// pcheckSubscription -
-//
-extern bool qAliasCompact(KjNode* qP, bool compact);
-static bool pcheckSubscription(ConnectionInfo* ciP, KjNode* subNodeP, bool idCanBePresent, KjNode** watchedAttributesPP, KjNode** timeIntervalPP, KjNode** qPP, KjNode** geoqPP)
-{
-  KjNode* idP                     = NULL;
-  KjNode* typeP                   = NULL;
-  KjNode* nameP                   = NULL;
-  KjNode* descriptionP            = NULL;
-  KjNode* entitiesP               = NULL;
-  KjNode* watchedAttributesP      = NULL;
-  KjNode* timeIntervalP           = NULL;
-  KjNode* qP                      = NULL;
-  KjNode* geoqP                   = NULL;
-  KjNode* csfP                    = NULL;
-  KjNode* isActiveP               = NULL;
-  KjNode* notificationP           = NULL;
-  KjNode* expiresP                = NULL;
-  KjNode* throttlingP             = NULL;
-  KjNode* temporalqP              = NULL;
-  int64_t dateTime;
-
-  if (subNodeP->type != KjObject)
-  {
-    orionldErrorResponseCreate(OrionldBadRequestData, "Invalid Subscription", "The payload data for updating a subscription must be a JSON Object");
-    ciP->httpStatusCode = SccBadRequest;
-    return false;
-  }
-
-  for (KjNode* nodeP = subNodeP->value.firstChildP; nodeP != NULL; nodeP = nodeP->next)
-  {
-    if (strcmp(nodeP->name, "id") == 0)
-    {
-      if (idCanBePresent == false)
-      {
-        orionldErrorResponseCreate(OrionldBadRequestData, "The Subscription ID cannot be modified", "id");
-        ciP->httpStatusCode = SccBadRequest;
-        return false;
-      }
-
-      DUPLICATE_CHECK(idP, "id", nodeP);
-      STRING_CHECK(nodeP, nodeP->name);
-      URI_CHECK(nodeP, nodeP->name);
-    }
-    else if (strcmp(nodeP->name, "type") == 0)
-    {
-      DUPLICATE_CHECK(typeP, "type", nodeP);
-      STRING_CHECK(nodeP, nodeP->name);
-
-      if (strcmp(nodeP->value.s, "Subscription") != 0)
-      {
-        orionldErrorResponseCreate(OrionldBadRequestData, "Invalid value for Subscription Type", nodeP->value.s);
-        ciP->httpStatusCode = SccBadRequest;
-        return false;
-      }
-    }
-    else if (strcmp(nodeP->name, "name") == 0)
-    {
-      DUPLICATE_CHECK(nameP, "name", nodeP);
-      STRING_CHECK(nodeP, nodeP->name);
-    }
-    else if (strcmp(nodeP->name, "description") == 0)
-    {
-      DUPLICATE_CHECK(descriptionP, "description", nodeP);
-      STRING_CHECK(nodeP, nodeP->name);
-    }
-    else if (strcmp(nodeP->name, "entities") == 0)
-    {
-      DUPLICATE_CHECK(entitiesP, "entities", nodeP);
-      if (pcheckEntities(ciP, entitiesP) == false)
-        return false;
-    }
-    else if (strcmp(nodeP->name, "watchedAttributes") == 0)
-    {
-      DUPLICATE_CHECK(watchedAttributesP, "watchedAttributes", nodeP);
-      ARRAY_CHECK(nodeP, nodeP->name);
-      for (KjNode* itemP = nodeP->value.firstChildP; itemP != NULL; itemP = itemP->next)
-      {
-        STRING_CHECK(itemP, "watchedAttributes item");
-        itemP->value.s = orionldContextItemExpand(orionldState.contextP, itemP->value.s, NULL, true, NULL);
-      }
-      *watchedAttributesPP = watchedAttributesP;
-    }
-    else if (strcmp(nodeP->name, "timeInterval") == 0)
-    {
-      DUPLICATE_CHECK(timeIntervalP, "timeInterval", nodeP);
-      INTEGER_CHECK(nodeP, "timeInterval");
-      *timeIntervalPP = timeIntervalP;
-    }
-    else if (strcmp(nodeP->name, "q") == 0)
-    {
-      DUPLICATE_CHECK(qP, "q", nodeP);
-      STRING_CHECK(nodeP, "q");
-      *qPP = qP;
-      LM_TMP(("QP: qP at %p", qP));
-      qAliasCompact(qP, false);
-    }
-    else if (strcmp(nodeP->name, "geoQ") == 0)
-    {
-      DUPLICATE_CHECK(geoqP, "geoQ", nodeP);
-      OBJECT_CHECK(nodeP, "geoQ");
-      if (pcheckGeoQ(ciP, nodeP) == false)
-        return false;
-      *geoqPP = geoqP;
-      LM_TMP(("QP: geoqP at %p", geoqP));
-    }
-    else if (strcmp(nodeP->name, "csf") == 0)
-    {
-      DUPLICATE_CHECK(csfP, "csf", nodeP);
-      STRING_CHECK(nodeP, "csf");
-    }
-    else if (strcmp(nodeP->name, "isActive") == 0)
-    {
-      DUPLICATE_CHECK(isActiveP, "isActive", nodeP);
-      BOOL_CHECK(nodeP, "isActive");
-    }
-    else if (strcmp(nodeP->name, "notification") == 0)
-    {
-      DUPLICATE_CHECK(notificationP, "notification", nodeP);
-      OBJECT_CHECK(nodeP, "notification");
-    }
-    else if (strcmp(nodeP->name, "expires") == 0)
-    {
-      DUPLICATE_CHECK(expiresP, "expires", nodeP);
-      STRING_CHECK(nodeP, "expires");
-      DATETIME_CHECK(expiresP->value.s, dateTime, "expires");
-    }
-    else if (strcmp(nodeP->name, "throttling") == 0)
-    {
-      DUPLICATE_CHECK(throttlingP, "throttling", nodeP);
-      INTEGER_CHECK(nodeP, "throttling");
-    }
-    else if (strcmp(nodeP->name, "temporalQ") == 0)
-    {
-      DUPLICATE_CHECK(temporalqP, "temporalQ", nodeP);
-      OBJECT_CHECK(nodeP, "temporalQ");
-    }
-    else if (strcmp(nodeP->name, "status") == 0)
-    {
-      orionldErrorResponseCreate(OrionldBadRequestData, "Attempt to modify Read-Only attribute", "status");
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-    else
-    {
-      LM_E(("Unknown field in Subscription fragment: '%s'", nodeP->name));
-      orionldErrorResponseCreate(OrionldBadRequestData, "Unknown field in Subscription fragment", nodeP->name);
-      ciP->httpStatusCode = SccBadRequest;
-      return false;
-    }
-  }
-
-  if ((notificationP != NULL) && (pcheckNotification(ciP, notificationP) == false))
-      return false;
-
-  return true;
-}
